[refactor] 전체 피드 조회의 내 띱 목록 조회 api의 통합 테스트 코드 수정 (#213)
